♻️ separate dialog wrapper and dialog content into different files
