+++ frontend/packages/erd-core/src/features/erd/components/ERDRenderer/CommandPalette/CommandPaletteContent/index.ts
@@ -1,0 +1,1 @@
+export * from './CommandPaletteContent'
